--- src/main/java/com/gatheringplatform/domain/BoardPlan.java
@@ -1,93 +1,0 @@
-package com.gatheringplatform.domain;
-
-import javax.validation.constraints.NotNull;
-
-public class BoardPlan {
-    @NotNull
-    private long id;
-    @NotNull
-    private String start_at ;
-    @NotNull
-    private String finish_at ;
-    @NotNull
-    private String frequency;
-    @NotNull
-    private int numOfPeople;
-    @NotNull
-    private String place;
-    private int fee;
-    private int min_age;
-    private int max_age;
-
-    public long getId() {
-        return id;
-    }
-
-    public void setId(long id) {
-        this.id = id;
-    }
-
-    public String getStart_at() {
-        return start_at;
-    }
-
-    public void setStart_at(String start_at) {
-        this.start_at = start_at;
-    }
-
-    public String getFinish_at() {
-        return finish_at;
-    }
-
-    public void setFinish_at(String finish_at) {
-        this.finish_at = finish_at;
-    }
-
-    public String getFrequency() {
-        return frequency;
-    }
-
-    public void setFrequency(String frequency) {
-        this.frequency = frequency;
-    }
-
-    public int getNumOfPeople() {
-        return numOfPeople;
-    }
-
-    public void setNumOfPeople(int numOfPeople) {
-        this.numOfPeople = numOfPeople;
-    }
-
-    public String getPlace() {
-        return place;
-    }
-
-    public void setPlace(String place) {
-        this.place = place;
-    }
-
-    public int getFee() {
-        return fee;
-    }
-
-    public void setFee(int fee) {
-        this.fee = fee;
-    }
-
-    public int getMin_age() {
-        return min_age;
-    }
-
-    public void setMin_age(int min_age) {
-        this.min_age = min_age;
-    }
-
-    public int getMax_age() {
-        return max_age;
-    }
-
-    public void setMax_age(int max_age) {
-        this.max_age = max_age;
-    }
-}
