+++ src/main/java/com/gatheringplatform/service/BoardService.java
@@ -1,7 +1,6 @@
 package com.gatheringplatform.service;
 
 import com.gatheringplatform.domain.Board;
-import com.gatheringplatform.domain.BoardPlan;
 import org.springframework.web.multipart.MultipartFile;
 
 import java.io.IOException;
@@ -10,5 +9,4 @@ import java.util.Map;
 public interface BoardService {
     public Map<String, String> uploadThumbNail(MultipartFile multipartFile) throws IOException;
     public String insertBoard(Board board) throws IOException;
-    public String insertPlan(BoardPlan boardPlan);
 }
